+++ docs/CHANGE_LOG.md
@@ -7,7 +7,7 @@ Change log for the Khipu app. This helps keep a running list of changes made for
 - Removed remaining unused functions from items.store.ts (theses were moved to categories-store.ts to support creating a list for add-category).
 - Added category selector with pre-fill (Issue #51).
 - Fixed the wikilist url created for the server call.
-- Using the fallacy functions to parse the cognitive bias Wikipedia page.
+- Using the table functions to parse the cognitive bias Wikipedia page.
 - Testing content with checkParent(item) && this.checkContent(item).
 
 ### Issues Closed